requirements.txt
kubernetes
flask
pyyaml
